--- a/core/libdeno.rs
+++ b/core/libdeno.rs
@@ -1,0 +1,1 @@
+../src/libdeno.rs