./target/debug/deno_core_test passes