@ry 2019-02-25
# this branch

~/src/deno2> ./third_party/wrk/mac/wrk -c 10 -d 30 http://127.0.0.1:8080/
Running 30s test @ http://127.0.0.1:8080/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   280.41us  466.28us  27.04ms   99.84%
    Req/Sec    17.71k     1.24k   19.98k    74.92%
  1060524 requests in 30.10s, 51.58MB read
Requests/sec:  35232.36
Transfer/sec:      1.71MB

# tools/node_tcp.js

~/src/deno2> ./third_party/wrk/mac/wrk -c 10 -d 30 http://127.0.0.1:4544/
Running 30s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   230.22us  109.64us   5.49ms   80.46%
    Req/Sec    21.12k     1.60k   26.48k    84.72%
  1264901 requests in 30.10s, 61.52MB read
Requests/sec:  42022.83
Transfer/sec:      2.04MB

# deno tests/http_bench.js

~/src/deno2> ./third_party/wrk/mac/wrk -c 10 -d 30 http://127.0.0.1:4500/
Running 30s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.77ms  739.31us  38.75ms   97.53%
    Req/Sec     2.90k   274.64     3.24k    94.33%
  173396 requests in 30.00s, 8.43MB read
Requests/sec:   5779.25
Transfer/sec:    287.83KB