@ry 2019-02-25
here's what strace looks like on a single request
https://gist.github.com/ry/4bcc56950b6fe2239c8d56fca5f23e38