@ry 2019-02-26
# current branch (osx, current_thread runtime)
```
~/src/deno2> ./third_party/wrk/mac/wrk -c 10 -d 10 http://127.0.0.1:8080/
Running 10s test @ http://127.0.0.1:8080/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   271.37us    0.87ms  30.20ms   98.59%
    Req/Sec    23.83k     2.05k   26.65k    87.62%
  478721 requests in 10.10s, 23.28MB read
Requests/sec:  47396.75
Transfer/sec:      2.31MB
```

# tools/node_tcp.js
```
~/src/deno2> ./third_party/wrk/mac/wrk -c 10 -d 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   266.51us  376.93us  16.44ms   98.92%
    Req/Sec    19.30k     2.25k   26.58k    81.68%
  388001 requests in 10.10s, 18.87MB read
Requests/sec:  38416.29
Transfer/sec:      1.87MB
```