@ry 2019-02-26
# linux, current_thread, f2a4784

## tools/node_tcp.js
```
> ./third_party/wrk/linux/wrk -c 10 -d 10 -t 1 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  1 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   146.45us  237.37us   8.44ms   98.05%
    Req/Sec    78.43k     9.08k   81.17k    97.00%
  779672 requests in 10.00s, 37.92MB read
Requests/sec:  77945.59
Transfer/sec:      3.79MB
```

## before "Respond to many async ops at once" (855b71910ddeab01c180e36b5114f7a5afb981d3)

```> ./third_party/wrk/linux/wrk -c 10 -d 10 -t 1 http://127.0.0.1:8080/
Running 10s test @ http://127.0.0.1:8080/
  1 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   232.57us  753.38us  25.71ms   97.03%
    Req/Sec    86.88k     7.31k   89.24k    97.00%
  864422 requests in 10.00s, 42.04MB read
Requests/sec:  86432.45
Transfer/sec:      4.20MB
```

## after "Respond to many async ops at once"
```
> ./third_party/wrk/linux/wrk -c 10 -d 10 -t 1 http://127.0.0.1:8080/
Running 10s test @ http://127.0.0.1:8080/
  1 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   137.95us  315.44us  14.34ms   97.17%
    Req/Sec    97.03k     3.27k   98.88k    92.00%
  965107 requests in 10.00s, 46.94MB read
Requests/sec:  96480.91
Transfer/sec:      4.69MB
```