@ry 2019-02-26
Now the benchmark supports true zero-copy buffers. Here are the benchmarks:

## With zero-copy buffer (7e7af40)
```
> ./third_party/wrk/linux/wrk -c 10 -d 30 http://127.0.0.1:8080/
Running 30s test @ http://127.0.0.1:8080/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   101.33us  260.48us  17.13ms   99.39%
    Req/Sec    57.52k     2.96k   62.40k    96.33%
  3433410 requests in 30.01s, 166.99MB read
Requests/sec: 114423.49
Transfer/sec:      5.57MB
```

## before zero-copy (62cd9ac806577399b316c2ad3a1812f0be9206b0)
```
> ./third_party/wrk/linux/wrk -c 10 -d 30 http://127.0.0.1:8080/
Running 30s test @ http://127.0.0.1:8080/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   143.05us  370.54us  17.98ms   98.84%
    Req/Sec    46.94k     2.95k   79.12k    94.34%
  2807129 requests in 30.10s, 136.53MB read
Requests/sec:  93261.26
Transfer/sec:      4.54MB
```

## node tools/http_tcp.js
```
> ./third_party/wrk/linux/wrk -c 10 -d 30 http://127.0.0.1:4544/
Running 30s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   144.88us  235.98us   9.57ms   98.89%
    Req/Sec    38.79k     2.60k   42.27k    98.17%
  2316103 requests in 30.01s, 112.65MB read
Requests/sec:  77179.39
Transfer/sec:      3.75MB
```