@ry 2019-02-26
@piscisaureus I think we should try to land this now. I hooked it up to benchmarks, so we can track further improvements there.